requirements.txt
cycler==0.10.0
decorator==4.4.0
imageio==2.5.0
imgaug>=0.2.8
kiwisolver==1.1.0
matplotlib>=3.0.3
networkx==2.3
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==6.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
PyWavelets==1.0.3
scikit-image>=0.15.0
scipy==1.2.1
Shapely==1.6.4.post1
six==1.12.0
torch>=1.0.0
torchvision==0.2.2.post3
